Brings cart to global scope
I have braught the cart from product scope to global scope wher cart
will be increment when any product is added to cart and will decrement
when any product is removed from cart. I also updated the functionality
of the cart from just the numbers of products being added to the array
of cart to which we can actully push the product id which is being added
and on removel from the cart it will search the id of the product being
removed from the cart and then remove it form cart array.